# Migrate nonzero from TH to ATen (CPU)