--- a/aten/src/ATen/native/TensorAdvancedIndexing.cpp
+++ b/aten/src/ATen/native/TensorAdvancedIndexing.cpp
@@ -1484,9 +1484,12 @@ Tensor& nonzero_out_cpu(const Tensor& self, Tensor& result) {
   const auto self_sizes = self.sizes();
   const auto total_nonzero = thread_count_nonzero.back();
   const int64_t ndim = self_sizes.size();
-  if (resize_output(result, {total_nonzero, ndim})) {
-    // Default to fortran-contiguous output (see gh-46224)
-    result.as_strided_({total_nonzero, ndim}, {1, total_nonzero});
+  const int64_t out_sizes[] = {total_nonzero, ndim};
+  if (resize_output_check(result, out_sizes)) {
+    // Reshape to fortran-contiguous output (see gh-46224)
+    auto impl = result.unsafeGetTensorImpl();
+    const int64_t out_strides[] = {1, total_nonzero};
+    at::native::resize_impl_cpu_(impl, out_sizes, out_strides);
   }
 
   if (result.numel() == 0) {