--- a/aten/src/TH/generic/THTensorMath.h
+++ b/aten/src/TH/generic/THTensorMath.h
@@ -4,7 +4,6 @@
 
 #include <ATen/core/Generator.h>
 
-TH_API void THTensor_(nonzero)(THLongTensor *subscript, THTensor *tensor);
 TH_API int THTensor_(equal)(THTensor *ta, THTensor *tb);
 
 #if !defined(TH_REAL_IS_HALF)