… aten/src/ATen/LegacyTHFunctionsCPU.cpp
… aten/src/TH/THTensorEvenMoreMath.cpp
… aten/src/TH/THTensorMath.cpp
… aten/src/TH/generic/THTensorEvenMoreMath.cpp
… aten/src/TH/generic/THTensorMath.cpp
… tools/build_variables.bzl