@peterbell10 2021-05-29
@JackCaoG any idea why these changes might be triggering XLA to segfault?