@ngimel 2021-05-29
If xla's implementation internally used count_nonzero, then changes to its dispatch could affect it, but I'm just guessing.