@ngimel 2021-05-30
Funny thing is, in pytorch's indexing implementation, when `nonzero` output is contiguous, then discontiguous slices would be used as indices https://github.com/pytorch/pytorch/blob/5c189946746599d47479c16f097dcc1beeecd7e0/aten/src/ATen/native/IndexingUtils.h#L43, and this PR fixes it and actually makes those slices contiguous. But I don't know what XLA implementation is doing.