@peterbell10 2021-06-01
@ngimel would it be reasonable to leave the output as contiguous for now and try to figure out the XLA issues in another PR.