@ngimel 2021-06-01
@peterbell10 yes, let's do that. 
@ailzhang, @JackCaoG we need `nonzero` to return a discontiguous tensor so that behavior is consistent with cuda, and indexing and sparse operations that depend on `nonzero` can get contiguous indices after `nonzero` return is chunked on the last dimension. XLA doesn't have strides, so it segfaults on attempts to allocate a discontiguous output tensor. What can we do to achieve the behavior we want on the cpu, while still keeping XLA happy?