@ailzhang 2021-06-02
@peterbell10 Thanks a lot for separating the change! If you send a PR to return non-contiguous output and cc me on it, I can take a look and see if there's anything we can do to unblock. Thanks!!