@leslie-fang-intel 2021-06-15
Hi @peterbell10. Thanks for the PR. I have one question, what will happen if I invoke the new nonzero in C++ with openmp.
For example:
```
# pragma omp parallel for simd schedule(static) if (omp_get_max_threads() > 1 && !omp_in_parallel())
for(int i=0;i<bs;i++){
   at::Tensor mask_index = at::nonzero(score > 0.05).squeeze(1);
}
```