@ngimel 2021-06-15
You cannot do this, `omp parallel` is the same in spirit as `parallel_for` #56794