@leslie-fang-intel 2021-06-15
Thanks for the comment.
Here is the case, I have a customized operations which will use `at::nonzero`
```
void my_customized_op(){
  at::nonzero();
}

at::parallel_for(0, numel, internal::GRAIN_SIZE, [&] (int64_t begin, int64_t end) {
  my_customized_op();
}
```
Openmp can't be nested. So It's expected that `at::nonzero` run with single thread in this case.
But I suspect in current implementation, `at::nonzero` will still create DimVector with unexpected num_threads sizes since it's not judge whether `at::nonzero` is in a nested openmp ENV or not.
https://github.com/pytorch/pytorch/blob/1c502d1f8ec861c31a08d580ae7b73b7fbebebed/aten/src/ATen/native/TensorAdvancedIndexing.cpp#L1549-L1550

And I am not sure: in this case what's the behavior of `at::nonzero`? I suspect I have got a different result but I may need more debug effort.