@ngimel 2021-06-15
According to the linked issue you cannot use tensor operations inside `parallel_for`, you example is not valid.