@leslie-fang-intel 2021-06-15
Sorry for the confusion. Actually I am using `OPENMP` for `my_customized_op`. 
```
void my_customized_op(){
  at::nonzero();
}

# pragma omp parallel for simd schedule(static) if (omp_get_max_threads() > 1 && !omp_in_parallel())
for(int i=0;i<bs;i++){
   my_customized_op();
}
```
It works previously. And I suspect the results got changed once we use the version of parallelized `at::nonzero`.