@peterbell10 2021-06-15
Is your question theoretical, or are you actually getting errors and/or wrong results?

Inside an existing `omp parallel` region, `parallel_for` should run single threaded. The single threaded case is known to work, since that's used for all inputs with fewer than ~32000 elements already. So, in theory I don't think there should be anything wrong here.

As for why, the `DimVector` is zero-initialized so it just sums the list `[num_nonzero, 0, 0, ...]` at the end, and there should be no problem at all.