@leslie-fang-intel 2021-06-17
Here are some thoughts for what I see in the UT with 2 threads: since we are using `lazy_init_num_threads`. https://github.com/pytorch/pytorch/blob/59b10036d557cf99df90a25d3a3a04d6312ac4f6/aten/src/ATen/ParallelOpenMP.h#L26
It will init the threads_number at the first time hit `parallel_for` inside a specific thread. So here are the cases:

- For the main thread, tid=0:

  Start running the `parallel_for`  inside the UT  => hit `lazy_init_num_threads` inside `parallel_for` and init the threads number as 2 => running `non_zero` op => create the `thread_begin` with size of 2 =>hit `parallel_for` inside `non_zero` again inside the `non_zero` op, but it will ignore `lazy_init_num_threads` since `init` is a thread_local var
  https://github.com/pytorch/pytorch/blob/59b10036d557cf99df90a25d3a3a04d6312ac4f6/aten/src/ATen/Parallel.h#L32
  So for the main thread, any time you invoke `at::get_num_threads()`, it will returns 2.


- For the second thread, tid=1:
 it will not do what's the main thread doing => Directly enter into `non_zero` op => create the `thread_begin` with size of 2 =>hit `parallel_for` inside `non_zero` again, init `lazy_init_num_threads` but here the `num_threads` is 1 now since we are inside a nested `openmp` region
https://github.com/pytorch/pytorch/blob/59b10036d557cf99df90a25d3a3a04d6312ac4f6/aten/src/ATen/ParallelOpenMP.cpp#L37
=> start running the `non_zero` in the second loop => create the `thread_begin` with size of 1, but the tid of current thread is still 1.

With some tests, if I move `lazy_init_num_threads` after https://github.com/pytorch/pytorch/blob/59b10036d557cf99df90a25d3a3a04d6312ac4f6/aten/src/ATen/ParallelOpenMP.h#L31-L34, it can workaround the issue we see.

CC @jgong5 for the discussion here.